+++ Tasks.Domain/Projects/Dtos/Works/ProjectWorkListDto.cs
@@ -10,6 +10,6 @@ namespace Tasks.Domain.Projects.Dtos.Works
         public DateTime EndTime { get; set; }
         public DeveloperListDto Developer { get; set; }
         public string Comment { get; set; }
-        public int Hours { get; set; }
+        public float Hours { get; set; }
     }
 }
